Create a file that do unit tests in algorithm file
I used your code as training object in my unit testing studying. This file is the result.
I have used the examples you wrote in the code and made positive tests of them. In the test of binary_convert I have written a exampel of tests for when incorrect values are entered into the functions. 

I hope the test is useful and you want to use it.

Thanks for the code. I wish you good luck in further developing it.